Update Transaction model to handle queries better
This commit changes the Transaction data model property 'txOutputs' so
that each entry in the array contains a more complete description of the
transaction.  Namely, we now include all of the following:

 - MoneyButton ID
 - Listoka User ID
 - Amount in USD transferred to that MoneyButton ID
 - A Comment ID if the transaction was to vote on a comment
 - A 'isListokaAcct' Boolean that can be set for outputs that are the
'service fees' of using the platform.

Additionally, this commit adds a 'raw' property on the Transaction model
that will save a stringified JSON object that is returned from the
MoneyButton service.  This should better future proof transactions -- in
the event that we need to change the model, we will have the actual
transaction data from the MoneyButton service to augment that
transformation.